% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{currency_summary}
\alias{currency_summary}
\title{Currency summary}
\usage{
currency_summary(from = "USD", to = "INR")
}
\arguments{
\item{from}{Currency to convert from}

\item{to}{Currency to convert to}
}
\description{
Contains information available via the Summary tab in Yahoo Finance
}
\examples{
currency_summary('GBP', 'USD')

}
