% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_api.R
\name{yf_get_financials}
\alias{yf_get_financials}
\title{Get Financial Statements}
\usage{
yf_get_financials(
  tickers,
  statement_type = c("income", "balance-sheet", "cash-flow")
)
}
\arguments{
\item{tickers}{Character vector of stock symbols.}

\item{statement_type}{The type of financial statement to retrieve. One of "income", "balance-sheet", or "cash-flow".}
}
\value{
A `tibble` of financial data with `symbol` as the first column.
}
\description{
Retrieves financial statements (income statement, balance sheet, or cash flow) for given tickers.
}
\examples{
\donttest{
yf_get_financials("AAPL", statement_type = "income")
}

}
