% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{currency_converter}
\alias{currency_converter}
\title{Currency converter}
\usage{
currency_converter(
  from = "EUR",
  to = "USD",
  start = NULL,
  end = NULL,
  period = "ytd",
  interval = "1d"
)
}
\arguments{
\item{from}{Currency to convert from}

\item{to}{Currency to convert to}

\item{start}{Specific starting date}

\item{end}{Specific ending date}

\item{period}{Length of time}

\item{interval}{Time between data points}
}
\description{
Retrieve current conversion rate between two currencies as well as historical rates
}
\examples{
currency_converter()
currency_converter('GBP', 'USD', '2022-07-01', '2022-07-10', '1h')

}
