% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{get_currencies}
\alias{get_currencies}
\title{Currencies}
\usage{
get_currencies()
}
\value{
Symbol, short and long name of the currencies.
}
\description{
List of currencies Yahoo Finance supports.
}
\examples{
\donttest{
get_currencies()
}

}
