% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{get_market_summary}
\alias{get_market_summary}
\title{Market Summary}
\usage{
get_market_summary(country = "US")
}
\arguments{
\item{country}{Name of the country.}
}
\value{
A \code{data.frame}.
}
\description{
Summary info of relevant exchanges for specific country.
}
\examples{
\donttest{
get_market_summary(country = 'US')
}

}
