% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{get_trending}
\alias{get_trending}
\title{Trending securities}
\usage{
get_trending(country = "US", count = 10)
}
\arguments{
\item{country}{Name of country}

\item{count}{Number of securities}
}
\description{
List of trending securities for specific country
}
\examples{
get_trending()

}
