% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indice.R
\docType{class}
\name{Indice-class}
\alias{Indice-class}
\alias{Index}
\title{R6 Class Representing an Index}
\format{
An R6 class object
}
\description{
Base class for getting all data related to indices from Yahoo Finance API.
}
\examples{

## ------------------------------------------------
## Method `Index$new`
## ------------------------------------------------

nifty_50 <- Index$new('^NSEI')

## ------------------------------------------------
## Method `Index$set_index`
## ------------------------------------------------

indice <- Index$new('^NSEI')
indice$set_index('^NDX')
}
\section{Super class}{
\code{\link[yahoofinancer:YahooFinanceBase]{yahoofinancer::YahooFinanceBase}} -> \code{Index}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{index}}{Deprecated. Returns \code{self$symbol}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Index-new}{\code{Index$new()}}
\item \href{#method-Index-set_index}{\code{Index$set_index()}}
\item \href{#method-Index-clone}{\code{Index$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="yahoofinancer" data-topic="YahooFinanceBase" data-id="get_history"><a href='../../yahoofinancer/html/YahooFinanceBase.html#method-YahooFinanceBase-get_history'><code>yahoofinancer::YahooFinanceBase$get_history()</code></a></span></li>
<li><span class="pkg-link" data-pkg="yahoofinancer" data-topic="YahooFinanceBase" data-id="set_symbol"><a href='../../yahoofinancer/html/YahooFinanceBase.html#method-YahooFinanceBase-set_symbol'><code>yahoofinancer::YahooFinanceBase$set_symbol()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Index-new"></a>}}
\if{latex}{\out{\hypertarget{method-Index-new}{}}}
\subsection{Method \code{new()}}{
Create a new Index object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Index$new(symbol = NA, index = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbol}}{Symbol}

\item{\code{index}}{Deprecated. Use \code{symbol} instead.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Index` object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{nifty_50 <- Index$new('^NSEI')
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Index-set_index"></a>}}
\if{latex}{\out{\hypertarget{method-Index-set_index}{}}}
\subsection{Method \code{set_index()}}{
Set a new index.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Index$set_index(symbol = NA, index = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbol}}{New symbol}

\item{\code{index}}{Deprecated. Use \code{symbol} instead.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{indice <- Index$new('^NSEI')
indice$set_index('^NDX')
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Index-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Index-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Index$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
