% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticker.R
\docType{class}
\name{Ticker-class}
\alias{Ticker-class}
\alias{Ticker}
\title{R6 Class Representing a Ticker}
\format{
An R6 class object
}
\description{
Class for retrieving all data related to a single ticker from the
Yahoo Finance API. Inherits \code{get_history()} from
\code{\link{YahooFinanceBase-class}}.
}
\seealso{
Other historical data: 
\code{\link{Index-class}},
\code{\link{Tickers}},
\code{\link{yf_download_prices}()},
\code{\link{yf_get_index_quotes}()}
}
\concept{historical data}
\section{Super class}{
\code{\link[yahoofinancer:YahooFinanceBase]{yahoofinancer::YahooFinanceBase}} -> \code{Ticker}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{valuation_measures}}{Retrieves valuation measures}

\item{\code{recommendations}}{Related symbols recommended by Yahoo Finance and their scores.}

\item{\code{technical_insights}}{Technical insights and indicators snapshot.}

\item{\code{currency}}{Currency code for the symbol.}

\item{\code{exchange_name}}{Short name of the stock exchange.}

\item{\code{full_exchange_name}}{Full name of the stock exchange.}

\item{\code{first_trade_date}}{Timestamp of the first recorded trade.}

\item{\code{regular_market_time}}{Timestamp of the last market trade.}

\item{\code{timezone}}{Timezone code of the exchange.}

\item{\code{exchange_timezone_name}}{Full timezone name of the exchange.}

\item{\code{regular_market_price}}{Current regular market price.}

\item{\code{fifty_two_week_high}}{52-week high price.}

\item{\code{fifty_two_week_low}}{52-week low price.}

\item{\code{regular_market_day_high}}{Highest price during the current trading session.}

\item{\code{regular_market_day_low}}{Lowest price during the current trading session.}

\item{\code{regular_market_volume}}{Current trading volume.}

\item{\code{previous_close}}{Closing price of the previous trading day.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Ticker-clone}{\code{Ticker$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="yahoofinancer" data-topic="YahooFinanceBase" data-id="get_history"><a href='../../yahoofinancer/html/YahooFinanceBase.html#method-YahooFinanceBase-get_history'><code>yahoofinancer::YahooFinanceBase$get_history()</code></a></span></li>
<li><span class="pkg-link" data-pkg="yahoofinancer" data-topic="YahooFinanceBase" data-id="initialize"><a href='../../yahoofinancer/html/YahooFinanceBase.html#method-YahooFinanceBase-initialize'><code>yahoofinancer::YahooFinanceBase$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="yahoofinancer" data-topic="YahooFinanceBase" data-id="set_symbol"><a href='../../yahoofinancer/html/YahooFinanceBase.html#method-YahooFinanceBase-set_symbol'><code>yahoofinancer::YahooFinanceBase$set_symbol()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ticker-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Ticker-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
