% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticker.R
\docType{class}
\name{Ticker-class}
\alias{Ticker-class}
\alias{Ticker}
\title{R6 Class Representing a Ticker}
\format{
An R6 class object
}
\description{
Base class for getting all data related to ticker from Yahoo Finance API
}
\examples{

## ------------------------------------------------
## Method `Ticker$new`
## ------------------------------------------------

aapl <- Ticker$new('aapl')

## ------------------------------------------------
## Method `Ticker$set_symbol`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
appl$set_symbol('msft')

## ------------------------------------------------
## Method `Ticker$get_asset_profile`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_asset_profile()

## ------------------------------------------------
## Method `Ticker$get_calendar_events`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_calendar_events()

## ------------------------------------------------
## Method `Ticker$get_company_officers`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_company_officers()

## ------------------------------------------------
## Method `Ticker$get_summary_profile`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_summary_profile()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{symbol}}{Symbol for which data is retrieved}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Ticker$new()}}
\item \href{#method-set_symbol}{\code{Ticker$set_symbol()}}
\item \href{#method-get_asset_profile}{\code{Ticker$get_asset_profile()}}
\item \href{#method-get_calendar_events}{\code{Ticker$get_calendar_events()}}
\item \href{#method-get_company_officers}{\code{Ticker$get_company_officers()}}
\item \href{#method-get_summary_profile}{\code{Ticker$get_summary_profile()}}
\item \href{#method-clone}{\code{Ticker$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Ticker object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$new(symbol = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbol}}{Symbol}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Ticker` object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_symbol"></a>}}
\if{latex}{\out{\hypertarget{method-set_symbol}{}}}
\subsection{Method \code{set_symbol()}}{
Set a new symbol.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$set_symbol(symbol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbol}}{New symbol}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
appl$set_symbol('msft')
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_asset_profile"></a>}}
\if{latex}{\out{\hypertarget{method-get_asset_profile}{}}}
\subsection{Method \code{get_asset_profile()}}{
Information related to the company's location, operations, and officers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$get_asset_profile()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
aapl$get_asset_profile()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_calendar_events"></a>}}
\if{latex}{\out{\hypertarget{method-get_calendar_events}{}}}
\subsection{Method \code{get_calendar_events()}}{
Earnings and Revenue expectations for upcoming earnings date.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$get_calendar_events()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
aapl$get_calendar_events()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_company_officers"></a>}}
\if{latex}{\out{\hypertarget{method-get_company_officers}{}}}
\subsection{Method \code{get_company_officers()}}{
Retrieves top executives for given symbol and their total pay package.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$get_company_officers()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
aapl$get_company_officers()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_summary_profile"></a>}}
\if{latex}{\out{\hypertarget{method-get_summary_profile}{}}}
\subsection{Method \code{get_summary_profile()}}{
Return business summary of given symbol
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$get_summary_profile()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
aapl$get_summary_profile()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
