% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticker.R
\docType{class}
\name{Ticker-class}
\alias{Ticker-class}
\alias{Ticker}
\title{R6 Class Representing a Ticker}
\format{
An R6 class object
}
\description{
Base class for getting all data related to ticker from Yahoo Finance API
}
\examples{

## ------------------------------------------------
## Method `Ticker$new`
## ------------------------------------------------

aapl <- Ticker$new('aapl')

## ------------------------------------------------
## Method `Ticker$set_symbol`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
appl$set_symbol('msft')

## ------------------------------------------------
## Method `Ticker$get_asset_profile`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_asset_profile()

## ------------------------------------------------
## Method `Ticker$get_calendar_events`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_calendar_events()

## ------------------------------------------------
## Method `Ticker$get_company_officers`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_company_officers()

## ------------------------------------------------
## Method `Ticker$get_earnings_history`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_earnings_history()

## ------------------------------------------------
## Method `Ticker$get_earnings`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_earnings()

## ------------------------------------------------
## Method `Ticker$get_earnings_trend`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_earnings_trend()

## ------------------------------------------------
## Method `Ticker$get_esg_scores`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_esg_scores()

## ------------------------------------------------
## Method `Ticker$get_financial_data`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_financial_data()

## ------------------------------------------------
## Method `Ticker$get_fund_bond_holdings`
## ------------------------------------------------

fund <- Ticker$new('vbmfx')
fund$get_fund_bond_holdings()

## ------------------------------------------------
## Method `Ticker$get_grading_history`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_grading_history()

## ------------------------------------------------
## Method `Ticker$get_index_trend`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_index_trend()

## ------------------------------------------------
## Method `Ticker$get_inside_holders`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_inside_holders()

## ------------------------------------------------
## Method `Ticker$get_insider_transactions`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_insider_transactions()

## ------------------------------------------------
## Method `Ticker$get_institution_ownership`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_institution_ownership()

## ------------------------------------------------
## Method `Ticker$get_major_holders`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_major_holders()

## ------------------------------------------------
## Method `Ticker$get_page_views`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_page_views()

## ------------------------------------------------
## Method `Ticker$get_price`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_price()

## ------------------------------------------------
## Method `Ticker$get_quote_type`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_quote_type()

## ------------------------------------------------
## Method `Ticker$get_recommendation_trend`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_recommendation_trend()

## ------------------------------------------------
## Method `Ticker$get_security_filings`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_security_filings()

## ------------------------------------------------
## Method `Ticker$get_share_purchase_activity`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_share_purchase_activity()

## ------------------------------------------------
## Method `Ticker$get_summary_detail`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_summary_detail()

## ------------------------------------------------
## Method `Ticker$get_summary_profile`
## ------------------------------------------------

aapl <- Ticker$new('aapl')
aapl$get_summary_profile()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{symbol}}{Symbol for which data is retrieved}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Ticker$new()}}
\item \href{#method-set_symbol}{\code{Ticker$set_symbol()}}
\item \href{#method-get_asset_profile}{\code{Ticker$get_asset_profile()}}
\item \href{#method-get_calendar_events}{\code{Ticker$get_calendar_events()}}
\item \href{#method-get_company_officers}{\code{Ticker$get_company_officers()}}
\item \href{#method-get_earnings_history}{\code{Ticker$get_earnings_history()}}
\item \href{#method-get_earnings}{\code{Ticker$get_earnings()}}
\item \href{#method-get_earnings_trend}{\code{Ticker$get_earnings_trend()}}
\item \href{#method-get_esg_scores}{\code{Ticker$get_esg_scores()}}
\item \href{#method-get_financial_data}{\code{Ticker$get_financial_data()}}
\item \href{#method-get_fund_bond_holdings}{\code{Ticker$get_fund_bond_holdings()}}
\item \href{#method-get_grading_history}{\code{Ticker$get_grading_history()}}
\item \href{#method-get_index_trend}{\code{Ticker$get_index_trend()}}
\item \href{#method-get_inside_holders}{\code{Ticker$get_inside_holders()}}
\item \href{#method-get_insider_transactions}{\code{Ticker$get_insider_transactions()}}
\item \href{#method-get_institution_ownership}{\code{Ticker$get_institution_ownership()}}
\item \href{#method-get_major_holders}{\code{Ticker$get_major_holders()}}
\item \href{#method-get_page_views}{\code{Ticker$get_page_views()}}
\item \href{#method-get_price}{\code{Ticker$get_price()}}
\item \href{#method-get_quote_type}{\code{Ticker$get_quote_type()}}
\item \href{#method-get_recommendation_trend}{\code{Ticker$get_recommendation_trend()}}
\item \href{#method-get_security_filings}{\code{Ticker$get_security_filings()}}
\item \href{#method-get_share_purchase_activity}{\code{Ticker$get_share_purchase_activity()}}
\item \href{#method-get_summary_detail}{\code{Ticker$get_summary_detail()}}
\item \href{#method-get_summary_profile}{\code{Ticker$get_summary_profile()}}
\item \href{#method-clone}{\code{Ticker$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Ticker object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$new(symbol = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbol}}{Symbol}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Ticker` object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_symbol"></a>}}
\if{latex}{\out{\hypertarget{method-set_symbol}{}}}
\subsection{Method \code{set_symbol()}}{
Set a new symbol.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$set_symbol(symbol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbol}}{New symbol}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
appl$set_symbol('msft')
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_asset_profile"></a>}}
\if{latex}{\out{\hypertarget{method-get_asset_profile}{}}}
\subsection{Method \code{get_asset_profile()}}{
Information related to the company's location, operations, and officers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$get_asset_profile()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
aapl$get_asset_profile()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_calendar_events"></a>}}
\if{latex}{\out{\hypertarget{method-get_calendar_events}{}}}
\subsection{Method \code{get_calendar_events()}}{
Earnings and Revenue expectations for upcoming earnings date.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$get_calendar_events()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
aapl$get_calendar_events()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_company_officers"></a>}}
\if{latex}{\out{\hypertarget{method-get_company_officers}{}}}
\subsection{Method \code{get_company_officers()}}{
Retrieves top executives for given symbol and their total pay package.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$get_company_officers()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
aapl$get_company_officers()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_earnings_history"></a>}}
\if{latex}{\out{\hypertarget{method-get_earnings_history}{}}}
\subsection{Method \code{get_earnings_history()}}{
Data related to historical earnings (actual vs. estimate)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$get_earnings_history()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
aapl$get_earnings_history()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_earnings"></a>}}
\if{latex}{\out{\hypertarget{method-get_earnings}{}}}
\subsection{Method \code{get_earnings()}}{
Historical earnings data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$get_earnings()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
aapl$get_earnings()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_earnings_trend"></a>}}
\if{latex}{\out{\hypertarget{method-get_earnings_trend}{}}}
\subsection{Method \code{get_earnings_trend()}}{
Historical trend data for earnings and revenue estimations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$get_earnings_trend()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
aapl$get_earnings_trend()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_esg_scores"></a>}}
\if{latex}{\out{\hypertarget{method-get_esg_scores}{}}}
\subsection{Method \code{get_esg_scores()}}{
Data related to environmental, social, and governance metrics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$get_esg_scores()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
aapl$get_esg_scores()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_financial_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_financial_data}{}}}
\subsection{Method \code{get_financial_data()}}{
Financial key performance indicators
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$get_financial_data()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
aapl$get_financial_data()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_fund_bond_holdings"></a>}}
\if{latex}{\out{\hypertarget{method-get_fund_bond_holdings}{}}}
\subsection{Method \code{get_fund_bond_holdings()}}{
Retrieves aggregated maturity and duration information for a given symbol
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$get_fund_bond_holdings()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fund <- Ticker$new('vbmfx')
fund$get_fund_bond_holdings()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_grading_history"></a>}}
\if{latex}{\out{\hypertarget{method-get_grading_history}{}}}
\subsection{Method \code{get_grading_history()}}{
Data related to upgrades / downgrades by companies
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$get_grading_history()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
aapl$get_grading_history()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_index_trend"></a>}}
\if{latex}{\out{\hypertarget{method-get_index_trend}{}}}
\subsection{Method \code{get_index_trend()}}{
Trend data related to given symbol's index, specificially PE and PEG ratios
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$get_index_trend()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
aapl$get_index_trend()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_inside_holders"></a>}}
\if{latex}{\out{\hypertarget{method-get_inside_holders}{}}}
\subsection{Method \code{get_inside_holders()}}{
Data related to stock holdings of a given symbol(s) insiders
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$get_inside_holders()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
aapl$get_inside_holders()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_insider_transactions"></a>}}
\if{latex}{\out{\hypertarget{method-get_insider_transactions}{}}}
\subsection{Method \code{get_insider_transactions()}}{
Transactions by insiders for a given symbol(s)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$get_insider_transactions()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
aapl$get_insider_transactions()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_institution_ownership"></a>}}
\if{latex}{\out{\hypertarget{method-get_institution_ownership}{}}}
\subsection{Method \code{get_institution_ownership()}}{
Top 10 owners of a given symbol
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$get_institution_ownership()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
aapl$get_institution_ownership()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_major_holders"></a>}}
\if{latex}{\out{\hypertarget{method-get_major_holders}{}}}
\subsection{Method \code{get_major_holders()}}{
Data showing breakdown of owners of given symbol(s), insiders, institutions, etc.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$get_major_holders()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
aapl$get_major_holders()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_page_views"></a>}}
\if{latex}{\out{\hypertarget{method-get_page_views}{}}}
\subsection{Method \code{get_page_views()}}{
Short, Mid, and Long-term trend data regarding a symbol's page views
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$get_page_views()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
aapl$get_page_views()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_price"></a>}}
\if{latex}{\out{\hypertarget{method-get_price}{}}}
\subsection{Method \code{get_price()}}{
Detailed pricing data for given symbol, exchange, quote type, currency, market cap, pre / post market data, etc.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$get_price()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
aapl$get_price()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_quote_type"></a>}}
\if{latex}{\out{\hypertarget{method-get_quote_type}{}}}
\subsection{Method \code{get_quote_type()}}{
Stock exchange specific data for given symbol
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$get_quote_type()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
aapl$get_quote_type()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_recommendation_trend"></a>}}
\if{latex}{\out{\hypertarget{method-get_recommendation_trend}{}}}
\subsection{Method \code{get_recommendation_trend()}}{
Data related to historical recommendations (buy, hold, sell) for a given symbol
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$get_recommendation_trend()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
aapl$get_recommendation_trend()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_security_filings"></a>}}
\if{latex}{\out{\hypertarget{method-get_security_filings}{}}}
\subsection{Method \code{get_security_filings()}}{
Historical SEC filings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$get_security_filings()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
aapl$get_security_filings()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_share_purchase_activity"></a>}}
\if{latex}{\out{\hypertarget{method-get_share_purchase_activity}{}}}
\subsection{Method \code{get_share_purchase_activity()}}{
High-level buy / sell data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$get_share_purchase_activity()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
aapl$get_share_purchase_activity()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_summary_detail"></a>}}
\if{latex}{\out{\hypertarget{method-get_summary_detail}{}}}
\subsection{Method \code{get_summary_detail()}}{
Contains information available via the Summary tab in Yahoo Finance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$get_summary_detail()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
aapl$get_summary_detail()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_summary_profile"></a>}}
\if{latex}{\out{\hypertarget{method-get_summary_profile}{}}}
\subsection{Method \code{get_summary_profile()}}{
Return business summary of given symbol
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$get_summary_profile()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{aapl <- Ticker$new('aapl')
aapl$get_summary_profile()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ticker$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
