% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate}
\alias{validate}
\title{Symbol validation}
\usage{
validate(symbol = NA, index = NA)
}
\arguments{
\item{symbol}{Ticker, index or fund name.}

\item{index}{Deprecated. Use \code{symbol} instead.}
}
\description{
Validate symbols before retrieving data.
}
\examples{
validate("aapl")
validate("aapls")

}
