% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\docType{class}
\name{YahooFinanceBase-class}
\alias{YahooFinanceBase-class}
\alias{YahooFinanceBase}
\title{R6 Class Representing a Yahoo Finance Base Object}
\format{
An R6 class object
}
\description{
Base class for Ticker and Index classes.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{symbol}}{Symbol for which data is retrieved.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-YahooFinanceBase-new}{\code{YahooFinanceBase$new()}}
\item \href{#method-YahooFinanceBase-set_symbol}{\code{YahooFinanceBase$set_symbol()}}
\item \href{#method-YahooFinanceBase-get_history}{\code{YahooFinanceBase$get_history()}}
\item \href{#method-YahooFinanceBase-clone}{\code{YahooFinanceBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-YahooFinanceBase-new"></a>}}
\if{latex}{\out{\hypertarget{method-YahooFinanceBase-new}{}}}
\subsection{Method \code{new()}}{
Create a new YahooFinanceBase object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{YahooFinanceBase$new(symbol = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbol}}{Symbol.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-YahooFinanceBase-set_symbol"></a>}}
\if{latex}{\out{\hypertarget{method-YahooFinanceBase-set_symbol}{}}}
\subsection{Method \code{set_symbol()}}{
Set a new symbol.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{YahooFinanceBase$set_symbol(symbol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbol}}{New symbol}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-YahooFinanceBase-get_history"></a>}}
\if{latex}{\out{\hypertarget{method-YahooFinanceBase-get_history}{}}}
\subsection{Method \code{get_history()}}{
Retrieves historical pricing data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{YahooFinanceBase$get_history(
  period = "ytd",
  interval = "1d",
  start = NULL,
  end = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{period}}{Length of time. Defaults to \code{'ytd'}.}

\item{\code{interval}}{Time between data points. Defaults to \code{'1d'}.}

\item{\code{start}}{Specific starting date. \code{String} or \code{date} object.}

\item{\code{end}}{Specific ending date. \code{String} or \code{date} object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.frame}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-YahooFinanceBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-YahooFinanceBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{YahooFinanceBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
