% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_api.R
\name{yf_get_index_quotes}
\alias{yf_get_index_quotes}
\title{Get Index Quotes}
\usage{
yf_get_index_quotes(index_symbol)
}
\arguments{
\item{index_symbol}{Character string of the index symbol (e.g., "^GSPC").}
}
\value{
A `tibble` containing the index quote data.
}
\description{
Retrieves current quotes and summary statistics for a given index symbol.
}
\examples{
\donttest{
yf_get_index_quotes("^GSPC")
}

}
